Consolidate granular UGens
